% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{classr-package}
\alias{classr}
\alias{classr-package}
\title{Course-Management Functions}
\description{
classr
}
\details{
Functions for Course Managment in the R Studio Environment.
}
\author{
Homer White (\email{hwhite0@georgetowncollege.edu}), Scott Switzer (\email{Scott_Switzer@georgetowncollege.edu})
}
\references{
\url{http://github.com/homerhanumat/classr}
}
\keyword{package}

