% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Collect}
\alias{Collect}
\title{Collect an Assignment from the Class}
\usage{
Collect(studentfile=NULL,dom=NULL,inst=NULL,assign=NULL,email=NULL)
}
\arguments{
\item{studentfile}{Text file containing student usernames, one per line.}

\item{dom}{the domain name}

\item{inst}{instructor's username}

\item{assign}{code for the assignment to be collected}

\item{email}{Receive an email confirming results?}
}
\value{
Side effects: graded copy returned to student, with permissions set so student can
open it.
}
\description{
Makes a copy of the ungraded assignment the student's submit folder and places it
in a subdirectory of your homework folder.
}
\examples{
\dontrun{
Collect(studentfile="students.txt",dom="WORK", inst="jdoe1",assign="HW01")
}
}
\author{
Scott Switzer and Homer White \email{hwhite0@georgetowncollege.edu}
}

