% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/complete_position.R
\name{complete_position}
\alias{complete_position}
\title{Enforce constant lenght position}
\usage{
complete_position(x, y)
}
\arguments{
\item{x}{Integer. Position.}

\item{y}{Integer. Number of characters}
}
\value{
Character. Position.
}
\description{
Add 0 to a number to force a constant length string.
}
\author{
Nicolas Mangin
}
