% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createDirectories}
\alias{createDirectories}
\title{Create Directories for the Class}
\usage{
createDirectories(studentfile=NULL,dom=NULL,key=NULL,passwordfile=NULL)
}
\arguments{
\item{studentfile}{Text file containing student usernames, one per line.}

\item{dom}{the domain name}

\item{key}{Your key to decrypt your password}

\item{passwordfile}{File containing your encrypted password.}
}
\value{
Side effects:  creation of directories.
}
\description{
Creates submit, returned and mynotes directories in students' Home folders.
}
\examples{
\dontrun{
createDirectories(studentfile="students.txt",dom="WORK",
     key="loveR",passwordfile="myLittleSecret.txt")
}
}
\author{
Scott Switzer and Homer White \email{hwhite0@georgetowncollege.edu}
}

