% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createInstance.r
\docType{methods}
\name{createInstance,character,ANY-method}
\alias{createInstance,character,ANY-method}
\title{Create Class Instance}
\usage{
\S4method{createInstance}{character,ANY}(cl, input = new.env(parent =
  emptyenv()), strict = FALSE, ...)
}
\arguments{
\item{cl}{\code{\link{character}}.}

\item{input}{\strong{Signature argument}.
Object containing object information that is relevant in the context
of instantiating.
This usually corresponds to an object that should be used "as is" as the
actual instance object and for which only the class path should be
updated.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if \code{cl} is not a class (checked by \code{isClass});
\code{FALSE}: no class existence check.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}

\item{cl}{\code{\link{ANY}}.}
}
\value{
\code{\link{ANY}}. Either \code{input} with modified class path or
   an object that is actually an empty \code{\link{environment}} but has an
   updated class path with class \code{cl} being the first element.
}
\description{
See generic: \code{\link[classr]{createInstance}}
}
\examples{
##------------------------------------------------------------------------------
## Default instance "body" //
##------------------------------------------------------------------------------

res <- createInstance("TestClass")
res

ls(res)
class(res)

##------------------------------------------------------------------------------
## Explicit object as instance "body" //
##------------------------------------------------------------------------------

res <- createInstance("TestClass", 
  input = list(x = TRUE, y = list(x_1 = 1, x_2 = 2)))
res 

## Strict //
try(createInstance("TestClass", strict = TRUE))
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/classr}
}
\seealso{
\code{
   \link[classr]{createInstance}
}
}

