% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/createInstance.r
\docType{methods}
\name{createInstance,character,missing-method}
\alias{createInstance,character,missing-method}
\title{Create Class Instance}
\usage{
\S4method{createInstance}{character,missing}(cl, input = new.env(parent =
  emptyenv()), strict = FALSE, ...)
}
\arguments{
\item{cl}{\code{\link{character}}.}

\item{input}{\code{\link{missing}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if \code{cl} is not a class (checked by \code{isClass});
\code{FALSE}: no class existence check.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link{createInstance-character-ANY}}.
}
\description{
See generic: \code{\link[classr]{createInstance}}
}
\examples{
##------------------------------------------------------------------------------
## Default instance "body" //
##------------------------------------------------------------------------------

res <- createInstance("TestClass")
res

ls(res)
class(res)

##------------------------------------------------------------------------------
## Explicit object as instance "body" //
##------------------------------------------------------------------------------

res <- createInstance("TestClass", 
  input = list(x = TRUE, y = list(x_1 = 1, x_2 = 2)))
res 

## Strict //
try(createInstance("TestClass", strict = TRUE))
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/rappster/classr}
}
\seealso{
\code{
   \link[classr]{createInstance}
}
}

