% Generated by roxygen2 (4.0.1): do not edit by hand
\name{createPasswordFile}
\alias{createPasswordFile}
\title{Create the Password File}
\usage{
createPasswordFile(password=NULL,file="myPasswordFile.txt",key=NULL)
}
\arguments{
\item{password}{Your password, as a charater string.}

\item{file}{name of the file in which to store the encrypted password.
Should reside in your Home directory.}

\item{key}{The key to be supplied to other functions to decrypt the password.}
}
\value{
Only the side effect:  the password file.
}
\description{
Creates a file conaining an encryption of your password.  The decyption is performed using a key
that you choose.  Clear your R history after using this function.
}
\examples{
\dontrun{
createPasswordFile(password="yoyo68peewee",file="myLittleSecret.txt",key="loveR")
}
}
\author{
Scott Switzer and Homer White \email{hwhite0@georgetowncollege.edu}
}

