% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jstree_to_tbltree.R
\name{jstree_to_tbltree}
\alias{jstree_to_tbltree}
\title{Transform a JSON tree into a table tree}
\usage{
jstree_to_tbltree(jstree, documents, document_types)
}
\arguments{
\item{jstree}{List. Document list as a classification tree.}

\item{documents}{Tibble.}

\item{document_types}{Tibble.}
}
\value{
Tibble. List of documents with positions extracted from the classification tree and additional variables gathered from the document list.
}
\description{
Function to transforms a document hierarchical classification into a tibble and update all document-related information
}
\author{
Nicolas Mangin
}
