% Generated by roxygen2 (4.0.1): do not edit by hand
\name{Return}
\alias{Return}
\title{Return an Assignment to the Class}
\usage{
Return(studentfile=NULL,dom=NULL,inst=NULL,assign=NULL,
           key=NULL,passwordfile=NULL,flag="_com")
}
\arguments{
\item{studentfile}{Text file containing student usernames, one per line.}

\item{dom}{the domain name}

\item{inst}{instructor's username}

\item{assign}{code for the assignment to be returned}

\item{key}{Your key to decrypt your password}

\item{passwordfile}{File containing your encrypted password.}

\item{flag}{the tag at the end of the file-name that identifies the copy with your comments.}
}
\value{
Side effects: graded copy returned to student, with permissions set so student can
open it.
}
\description{
Picks the graded, commented copy of the assignment from the student's folder in your
homework directory and places it the student's returned directory.
}
\examples{
\dontrun{
Return(studentfile="students.txt",dom="WORK", inst="jdoe1",assign="HW01",
     key="loveR",passwordfile="myLittleSecret.txt",flag="_com")
}
}
\author{
Scott Switzer and Homer White \email{hwhite0@georgetowncollege.edu}
}

