% Generated by roxygen2 (4.0.1): do not edit by hand
\name{SubmissionReport}
\alias{SubmissionReport}
\title{Turn-In Report}
\usage{
SubmissionReport(studentfile=NULL,folder=NULL,assigntype=NULL)
}
\arguments{
\item{studentfile}{Text file containing student usernames, one per line.}

\item{folder}{path name to folder in Home directory that contains assignments of the desired type}

\item{assigntype}{substring of directory-name, indicating type of assignment}
}
\value{
a data frame
}
\description{
Makes a data frame showing which students have turned in assignments of a given type.
}
\examples{
\dontrun{
SubmissionReport(studentfile="students.txt", folder="homework/mat331" ,assigntype="HW")
}
}
\author{
Homer White \email{hwhite0@georgetowncollege.edu}
}

