% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_add.R
\name{tags_add}
\alias{tags_add}
\title{Add a tag}
\usage{
tags_add(course_paths, tag_label)
}
\arguments{
\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{tag_label}{Character. Label of the tag which should be removed.}
}
\value{
Add an empty tag in all documents.
}
\description{
Function adding a tag to all documents.
}
\author{
Nicolas Mangin
}
