% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_batch_change_server.R
\name{tags_batch_change_server}
\alias{tags_batch_change_server}
\title{Batch changes tags in documents}
\usage{
tags_batch_change_server(id, tags, documents)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{tags}{Tibble. List of tags, their values and attributes.}

\item{documents}{Tibble. List of documents}
}
\description{
Interface allowing the user to apply changes to tags in all documents.
}
\author{
Nicolas Mangin
}
