% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_batch_change_ui.R
\name{tags_batch_change_ui}
\alias{tags_batch_change_ui}
\title{Batch changes tags in documents}
\usage{
tags_batch_change_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Interface facilitating the modification of tag labels or values in all documents.
}
\description{
Interface allowing the user to apply changes to tags in all documents.
}
\author{
Nicolas Mangin
}
