% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_edit_server.R
\name{tags_edit_server}
\alias{tags_edit_server}
\title{Edit tags}
\usage{
tags_edit_server(id, tags, documents)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{tags}{Tibble. List of tags, their values and attributes.}

\item{documents}{Tibble. List of documents}
}
\value{
Modified tags.
}
\description{
Tabular form to change the attributes associated with tags or add new tags.
}
\author{
Nicolas Mangin
}
