% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_edit_ui.R
\name{tags_edit_ui}
\alias{tags_edit_ui}
\title{Edit tags}
\usage{
tags_edit_ui(id)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}
}
\value{
Modified tags.
}
\description{
Tabular form to change the attributes associated with tags or add new tags.
}
\author{
Nicolas Mangin
}
