% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_remove.R
\name{tags_remove}
\alias{tags_remove}
\title{Remove a tag from all documents}
\usage{
tags_remove(tags, documents, tag_label)
}
\arguments{
\item{tags}{Tibble. List of tags, their values and attributes.}

\item{documents}{Tibble. List of documents}

\item{tag_label}{Character. Label of the tag which should be removed.}
}
\value{
Remove the specified tag from all documents.
}
\description{
Function removing a tag from all documents.
}
\author{
Nicolas Mangin
}
