% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_replace_value.R
\name{tags_replace_value}
\alias{tags_replace_value}
\title{Replace tag value}
\usage{
tags_replace_value(tags, documents, tag_label, old_value, new_value)
}
\arguments{
\item{tags}{Tibble. List of tags, their values and attributes.}

\item{documents}{Tibble. List of documents}

\item{tag_label}{Character. Label of the tag in which the value should be replaced.}

\item{old_value}{Character. Value which should be replaced.}

\item{new_value}{Character. Value replacing the old value (should exist in the tag list before).}
}
\value{
Apply tag value change in all documents.
}
\description{
Function replacing the value of a tag by another in all documents.
}
\author{
Nicolas Mangin
}
