% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_search_icons_ui.R
\name{tags_search_icons_ui}
\alias{tags_search_icons_ui}
\title{Search icons}
\usage{
tags_search_icons_ui(id)
}
\arguments{
\item{id}{ID if the module to connect the user interface to the appropriate server side.}
}
\description{
Module allowing the user to search fontawesome icons to illustrate tags.
}
\author{
Nicolas Mangin
}
