% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbltree_to_jstree.R
\name{tbltree_to_jstree}
\alias{tbltree_to_jstree}
\title{Transform a table tree to a json tree}
\usage{
tbltree_to_jstree(tbltree)
}
\arguments{
\item{tbltree}{Tibble. Classification tree for documents.}
}
\value{
List interpreted as a JSON file by jsTreeR
}
\description{
Function transforming a table containing a list of documents (with a position in a hierarchy) into a json file structure. Note that the classification assumes three roots: Included, Excluded and Unclassified.
}
\author{
Nicolas Mangin
}
