% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_find_higher_level.R
\name{tree_find_higher_level}
\alias{tree_find_higher_level}
\title{Find the higher level folder}
\usage{
tree_find_higher_level(x, y)
}
\arguments{
\item{x}{Character. Level of the document.}

\item{y}{Character. Higher-level position which should be returned.}
}
\value{
Character. Higher-level position
}
\description{
Function returning the required higher-level position of a document in a tree.
}
\author{
Nicolas Mangin
}
