% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_edit_server.R
\name{trees_edit_server}
\alias{trees_edit_server}
\title{Edit course tree}
\usage{
trees_edit_server(id, intake, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}

\item{tree}{Reactive. Function containing a list of documents as a classification tree compatible with jsTreeR.}
}
\value{
Save the tree tree as a tibble in the folder containing trees.
}
\description{
Module allowing the user to change the hierarchical classification (tree) of documents and save these changes, either with the same name or a different one.
}
\author{
Nicolas Mangin
}
