% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_edit_ui.R
\name{trees_edit_ui}
\alias{trees_edit_ui}
\title{Edit course tree}
\usage{
trees_edit_ui(id)
}
\arguments{
\item{id}{ID if the module to connect the user interface to the appropriate server side.}
}
\value{
Save the tree tree as a tibble in the folder containing trees.
}
\description{
Module allowing the user to change the hierarchical classification (tree) of documents and save these changes, either with the same name or a different one.
}
\author{
Nicolas Mangin
}
