% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_fill_list_level.R
\name{trees_fill_list_level}
\alias{trees_fill_list_level}
\title{Transform a tibble to a json}
\usage{
trees_fill_list_level(root = NULL, children = NULL, text = NULL)
}
\arguments{
\item{root}{Tibble.}

\item{children}{Tibble.}

\item{text}{Character.}
}
\value{
List interpreted as a JSON file by jsTreeR
}
\description{
Function transforming a table containing a list of documents (with a position in a hierarchy) into a json file structure. Note that the classification assumes three roots: Included, Excluded and Unclassified.
}
\seealso{
trees_tibble_to_json

trees_create_list_level
}
\author{
Nicolas Mangin
}
