% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_json_to_tibble.R
\name{trees_json_to_tibble}
\alias{trees_json_to_tibble}
\title{Transform a JSON list into a tibble}
\usage{
trees_json_to_tibble(json_file, maxlevel = 10, course_data)
}
\arguments{
\item{json_file}{List. Document list as a classification tree.}

\item{maxlevel}{Maximum number of nested levels to be expected. Set to a high number to give leeway}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}
}
\value{
Tibble. List of documents with positions extracted from the classification tree and additional variables gathered from the document list.
}
\description{
Function to transforms a document classification into a tibble and update all document-related information
}
\author{
Nicolas Mangin
}
