% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_selected_json_to_tibble.R
\name{trees_selected_json_to_tibble}
\alias{trees_selected_json_to_tibble}
\title{Tree selection as tibble}
\usage{
trees_selected_json_to_tibble(selected_tree)
}
\arguments{
\item{selected_tree}{List. input from a jsTreeR output.}
}
\value{
Tibble. List of selected documents.
}
\description{
Function transforming a selection of documents from jsTreeR to a table
}
\author{
Nicolas Mangin
}
