% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_structure_textbook.R
\name{trees_structure_textbook}
\alias{trees_structure_textbook}
\title{Create a textbook structure}
\usage{
trees_structure_textbook(tree, tree_name = "", website = "")
}
\arguments{
\item{tree}{Tibble.}

\item{tree_name}{Character. Name of the tree.}

\item{website}{Character. Address of the website where the textbook is hosted.}
}
\value{
Tibble with all necessary information about the pages of the textbook.
}
\description{
Function creating the textbook structure associated with a tree.
}
\author{
Nicolas Mangin
}
