% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trees_tibble_to_json.R
\name{trees_tibble_to_json}
\alias{trees_tibble_to_json}
\title{Transform a tibble to a json}
\usage{
trees_tibble_to_json(tree_tibble)
}
\arguments{
\item{tree_tibble}{Tibble. Classification tree for documents.}
}
\value{
List interpreted as a JSON file by jsTreeR
}
\description{
Function transforming a table containing a list of documents (with a position in a hierarchy) into a json file structure. Note that the classification assumes three roots: Included, Excluded and Unclassified.
}
\seealso{
trees_create_list_levels

trees_fill_list_level
}
\author{
Nicolas Mangin
}
