% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{methods}
\name{createInstance,character,missing-method}
\alias{createInstance,character,missing-method}
\title{Create Class Instance}
\usage{
\S4method{createInstance}{character,missing}(cl, obj = new.env(parent =
  emptyenv()), strict = FALSE, ...)
}
\arguments{
\item{cl}{\code{\link{character}}.}

\item{obj}{\code{\link{missing}}.}

\item{strict}{\code{\link{logical}}.
\code{TRUE}: error if \code{cl} is not a class (checked by \code{isClass});
\code{FALSE}: no class existence check.}

\item{...}{Further arguments to be passed to subsequent functions/methods.}
}
\value{
See method
   \code{\link{createInstance-character-ANY}}.
}
\description{
See generic: \code{\link[classr]{createInstance}}
}
\examples{
\dontrun{

## Default instance "body" //
res <- createInstance(cl = "TestClass")
res

ls(res)
class(res)

## Explicit object as instance "body" //
res <- createInstance(cl = "TestClass", obj = list(x = TRUE))
res 

## Strict //
try(createInstance(cl = "TestClass", strict = TRUE))

}
}
\author{
Janko Thyson \email{janko.thyson@rappster.de}
}
\references{
\url{http://github.com/Rappster/classr}
}
\seealso{
\code{
   \link[classr]{createInstance}
}
}

