% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_add.R
\name{tags_add}
\alias{tags_add}
\title{Add a tag}
\usage{
tags_add(tags, documents, tag_label)
}
\arguments{
\item{tags}{Tibble. List of tags, their values and attributes.}

\item{documents}{Tibble. List of documents}

\item{tag_label}{Character. Label of the tag which should be added.}
}
\value{
Add an empty tag in all documents.
}
\description{
Function adding a tag to all documents.
}
\author{
Nicolas Mangin
}
