% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags_edit_server.R
\name{tags_edit_server}
\alias{tags_edit_server}
\title{Edit tags}
\usage{
tags_edit_server(id, course_data, course_paths)
}
\arguments{
\item{id}{Character. ID of the module to connect the user interface to the appropriate server side.}

\item{course_data}{Reactive. Function containing all the course data loaded with the course.}

\item{course_paths}{Reactive. Function containing a list of paths to the different folders and databases on local disk.}
}
\value{
Modified tags.
}
\description{
Tabular form to change the attributes associated with tags or add new tags.
}
\author{
Nicolas Mangin
}
